Performance and bugs fixed

+++ api/controllers.py
@@ -643,18 +643,22 @@ def delete_google_drive_backup(user_id):
     google_integration = GoogleDriveIntegrationDB()
     token_db = Oauth_tokens_db()
     oauth_tokens = token_db.get_by_user_id(user_id)
-    
-    if google_integration.get_by_user_id(user_id) is None:
-        google_integration.create(user_id, 0)
+    google_drive_option =google_integration.get_by_user_id(user_id) 
+    if google_drive_option == None:
+        return {"message": "Google drive sync is not enabled"}, 403
+    if google_drive_option.isEnabled == 0:
+        return {"message": "Google drive sync is not enabled"}, 403
     if not oauth_tokens:
-        GoogleDriveIntegrationDB().update_google_drive_sync(user_id, 0)
-        return {"message": "Google drive sync is not enabled"}, 200
+        return {"message": "Google drive sync is not enabled"}, 403
     sse = ServiceSideEncryption()
     try:
         creds_b64 = sse.decrypt( ciphertext=oauth_tokens.enc_credentials, nonce=oauth_tokens.cipher_nonce,  tag=oauth_tokens.cipher_tag)
         credentials = json.loads(base64.b64decode(creds_b64).decode("utf-8"))
-        google_drive_api.delete_all_backups(credentials=credentials)
-        return {"message": "Backups deleted"}, 200
+        status = google_drive_api.delete_all_backups(credentials=credentials)
+        if status :
+            return {"message": "Backups deleted"}, 200
+        else:
+            return {"message": "Error while deleting backups"}, 500
     except Exception as e:
         logging.error("Error while deleting backup from google drive " + str(e))
         token_db.delete(user_id)
